refactor(security): consolidate EnhancedSecurityLayer into SecurityLayer (#10666)
Fold all unique features from EnhancedSecurityLayer (command execution,
approval workflow, audit log, sandbox, pending-approvals dict) into the
canonical SecurityLayer class. Delete enhanced_security_layer.py and migrate
all callers: api/security.py, security/{intelligent_agent_security,
secure_llm_command_parser}.py, and all related tests.

Security checks preserved verbatim — role-downgrade guard, command risk
assessment, approval timeout, single-user mode warning, auth-default True.

Part of #10666
