--- autobot-backend/enhanced_security_layer.py
@@ -1,627 +1,0 @@
-# Copyright 2025-2026 mrveiss
-# SPDX-License-Identifier: Apache-2.0
-# AutoBot - AI-Powered Automation Platform
-# Author: mrveiss
-"""
-Enhanced Security Layer with Command Execution Controls
-Integrates secure command execution with role-based permissions
-"""
-
-import asyncio
-import datetime
-import json
-import logging
-import os
-from datetime import timezone
-from typing import Any, Dict, List
-
-from autobot_shared.async_compat import run_or_schedule
-from autobot_shared.logging_manager import get_logger
-from autobot_shared.ssot_config import config
-
-# Import the centralized ConfigManager
-from config import config as global_config_manager
-from secure_command_executor import CommandRisk, SecureCommandExecutor, SecurityPolicy
-
-logger = get_logger(__name__)
-
-# Audit log file path resolved from AUTOBOT_AUDIT_LOG_FILE env var at import time.
-# Set AUTOBOT_AUDIT_LOG_FILE to override the default path.
-_AUDIT_LOG_FILE_DEFAULT = "/opt/autobot/logs/audit.log"
-
-
-def _resolve_audit_log_file() -> str:
-    """Return audit log file path from AUTOBOT_AUDIT_LOG_FILE env var with logged fallback."""
-    value = config.audit_log_file
-    if not value:
-        logger.warning(
-            "AUTOBOT_AUDIT_LOG_FILE is not set or empty; falling back to %s",
-            _AUDIT_LOG_FILE_DEFAULT,
-        )
-        return _AUDIT_LOG_FILE_DEFAULT
-    return value
-
-
-_AUDIT_LOG_FILE = _resolve_audit_log_file()
-
-# Performance optimization: O(1) lookup for security checks (Issue #326)
-DEPRECATED_PRIVILEGED_ROLES = {"god", "superuser", "root"}
-HIGH_RISK_COMMAND_RISKS = {CommandRisk.HIGH, CommandRisk.MODERATE}
-COMMAND_EXECUTION_ACTIONS = {"command_execution_attempt", "command_execution_complete"}
-
-
-def _parse_audit_log_entry(line: str, user: str | None = None) -> Dict[str, Any] | None:
-    """Parse a single audit log line and filter by user (Issue #315: extracted).
-
-    Args:
-        line: JSON line from audit log
-        user: Optional user filter
-
-    Returns:
-        Parsed entry dict if valid command execution, None otherwise
-    """
-    try:
-        entry = json.loads(line)
-    except json.JSONDecodeError:
-        return None
-
-    if "action" not in entry:
-        return None
-    if entry["action"] not in COMMAND_EXECUTION_ACTIONS:
-        return None
-    if user is not None and entry.get("user") != user:
-        return None
-
-    return entry
-
-
-class EnhancedSecurityLayer:
-    """
-    Enhanced security layer that integrates command execution security
-    with role-based access control and audit logging
-    """
-
-    def __init__(self):
-        """Initialize security layer with RBAC, auditing, and command executor."""
-        # Use centralized config manager
-        self.security_config = global_config_manager.get("security_config", {})
-        self.enable_auth = self.security_config.get("enable_auth", False)
-        self.enable_command_security = self.security_config.get("enable_command_security", True)
-        self.audit_log_file = self.security_config.get("audit_log_file") or _AUDIT_LOG_FILE
-        self.roles = self.security_config.get("roles", {})
-        self.allowed_users = self.security_config.get("allowed_users", {})
-
-        # Command security settings
-        self.command_approval_required = self.security_config.get("command_approval_required", True)
-        self.use_docker_sandbox = self.security_config.get("use_docker_sandbox", False)
-
-        # Initialize secure command executor
-        self.command_executor = SecureCommandExecutor(
-            policy=self._create_security_policy(),
-            require_approval_callback=self._command_approval_callback,
-            use_docker_sandbox=self.use_docker_sandbox,
-        )
-
-        # Initialize enhanced sandbox executor if Docker is available
-        self.sandbox_executor = None
-        if self.use_docker_sandbox:
-            try:
-                from secure_sandbox_executor import secure_sandbox
-
-                self.sandbox_executor = secure_sandbox
-                logger.info("Enhanced Docker sandbox executor initialized")
-            except Exception as e:
-                logger.warning("Failed to initialize enhanced sandbox executor: %s", e)
-
-        # Approval queue for async command approvals
-        self.pending_approvals: Dict[str, asyncio.Event] = {}
-        self.approval_results: Dict[str, bool] = {}
-
-        if self.audit_log_file:
-            log_dir = os.path.dirname(self.audit_log_file)
-            if log_dir:
-                os.makedirs(log_dir, exist_ok=True)
-        logger.info("EnhancedSecurityLayer initialized")
-        logger.debug("Authentication enabled: %s", self.enable_auth)
-        logger.debug("Command security enabled: %s", self.enable_command_security)
-        logger.debug("Docker sandbox: %s", self.use_docker_sandbox)
-        logger.debug("Audit log file: %s", self.audit_log_file)
-
-    def _create_security_policy(self) -> SecurityPolicy:
-        """Create security policy from configuration"""
-        policy = SecurityPolicy()
-
-        # Load custom policies from config if available
-        custom_policies = self.security_config.get("command_policies", {})
-
-        # #7161: SAFE_COMMANDS / FORBIDDEN_COMMANDS are frozensets after the
-        # #765 centralized-command-patterns refactor. The previous .update()
-        # calls crashed with AttributeError on any deployment that supplied
-        # `command_policies` in config — silently broken because no test
-        # exercised this path until #6987 fixed the src.* mocks. Now we
-        # rebuild the attribute as a mutable set merging defaults + customs.
-        if "safe_commands" in custom_policies:
-            policy.safe_commands = set(policy.safe_commands) | set(custom_policies["safe_commands"])
-
-        if "forbidden_commands" in custom_policies:
-            policy.forbidden_commands = set(policy.forbidden_commands) | set(custom_policies["forbidden_commands"])
-
-        if "allowed_paths" in custom_policies:
-            from pathlib import Path
-
-            policy.allowed_paths = [Path(p) for p in custom_policies["allowed_paths"]]
-
-        return policy
-
-    def _log_approval_request(self, command_id: str, approval_data: Dict[str, Any]) -> None:
-        """Log command approval request to audit log. Issue #620."""
-        self.audit_log(
-            action="command_approval_request",
-            user="system",
-            outcome="pending",
-            details={
-                "command": approval_data["command"],
-                "risk": approval_data["risk"],
-                "reasons": approval_data["reasons"],
-                "command_id": command_id,
-            },
-        )
-
-    def _check_auto_approve_moderate(self, command_id: str, approval_data: Dict[str, Any]) -> bool | None:
-        """Check if moderate risk command should be auto-approved. Issue #620.
-
-        Returns:
-            True if auto-approved, None if approval still required.
-        """
-        if approval_data["risk"] != CommandRisk.MODERATE.value:
-            return None
-        if not self.security_config.get("auto_approve_moderate", False):
-            return None
-        self.audit_log(
-            action="command_auto_approved",
-            user="system",
-            outcome="approved",
-            details={"command_id": command_id, "risk": "moderate"},
-        )
-        return True
-
-    async def _wait_for_approval_response(self, command_id: str) -> bool:
-        """Wait for approval response with timeout and log result. Issue #620.
-
-        Returns:
-            True if approved, False if denied or timeout.
-        """
-        approval_event = asyncio.Event()
-        self.pending_approvals[command_id] = approval_event
-
-        try:
-            await asyncio.wait_for(approval_event.wait(), timeout=300)
-            approved = self.approval_results.get(command_id, False)
-            self.audit_log(
-                action="command_approval_response",
-                user="system",
-                outcome="approved" if approved else "denied",
-                details={"command_id": command_id},
-            )
-            return approved
-        except asyncio.TimeoutError:
-            self.audit_log(
-                action="command_approval_timeout",
-                user="system",
-                outcome="denied",
-                details={"command_id": command_id},
-            )
-            return False
-        finally:
-            self.pending_approvals.pop(command_id, None)
-            self.approval_results.pop(command_id, None)
-
-    async def _command_approval_callback(self, approval_data: Dict[str, Any]) -> bool:
-        """
-        Callback for command approval requests.
-        This can be extended to integrate with UI or notification systems.
-        """
-        command_id = f"cmd_{int(approval_data['timestamp'])}"
-        self._log_approval_request(command_id, approval_data)
-
-        # Check for auto-approval of moderate risk commands
-        auto_result = self._check_auto_approve_moderate(command_id, approval_data)
-        if auto_result is not None:
-            return auto_result
-
-        # Wait for manual approval
-        return await self._wait_for_approval_response(command_id)
-
-    def approve_command(self, command_id: str, approved: bool = True):
-        """
-        Approve or deny a pending command
-        This would be called from UI or API endpoint
-        """
-        if command_id in self.pending_approvals:
-            self.approval_results[command_id] = approved
-            self.pending_approvals[command_id].set()
-
-    def _handle_deprecated_role(self, user_role: str, action_type: str, resource: str | None) -> str:
-        """Handle deprecated privileged roles by logging and downgrading. Issue #620.
-
-        Args:
-            user_role: Original user role
-            action_type: Action being attempted
-            resource: Resource being accessed
-
-        Returns:
-            Downgraded role name (admin) for deprecated roles
-        """
-        self.audit_log(
-            action="deprecated_role_usage",
-            user=user_role,
-            outcome="warning",
-            details={
-                "deprecated_role": user_role,
-                "action_attempted": action_type,
-                "resource": resource,
-                "message": "God/superuser/root roles deprecated. Update to admin role.",
-            },
-        )
-        return "admin"
-
-    def _check_shell_execute_permission(self, role_permissions: List[str]) -> bool:
-        """Check if role has shell execution permission. Issue #620.
-
-        Args:
-            role_permissions: List of permissions for the role
-
-        Returns:
-            True if shell execution is allowed
-        """
-        if "allow_shell_execute" in role_permissions:
-            return True
-        if "allow_shell_execute_safe" in role_permissions:
-            return True
-        return False
-
-    def _check_wildcard_permissions(self, action_type: str, role_permissions: List[str]) -> bool:
-        """Check if action matches any wildcard permissions. Issue #620.
-
-        Args:
-            action_type: The action to check
-            role_permissions: List of permissions for the role
-
-        Returns:
-            True if action matches a wildcard permission
-        """
-        for permission in role_permissions:
-            if permission.endswith(".*"):
-                permission_prefix = permission[:-1]
-                if action_type.startswith(permission_prefix):
-                    return True
-        return False
-
-    def check_permission(self, user_role: str, action_type: str, resource: str | None = None) -> bool:
-        """
-        Enhanced permission checking that includes command execution permissions
-        SECURITY FIX: Removed god mode bypass - all roles use granular RBAC
-        """
-        if not self.enable_auth:
-            return True
-
-        # Handle deprecated roles (O(1) lookup - Issue #326)
-        if user_role.lower() in DEPRECATED_PRIVILEGED_ROLES:
-            user_role = self._handle_deprecated_role(user_role, action_type, resource)
-
-        role_permissions = self.roles.get(user_role, {}).get("permissions", [])
-
-        # #7161: previously the shell_execute special-case checked ONLY
-        # configured `self.roles` permissions, ignoring `_get_default_role_permissions()`.
-        # Roles like 'admin' that have `allow_shell_execute` only in defaults
-        # (not configured) were denied shell access. Now combine both sources
-        # so the special-case respects defaults too.
-        default_permissions = self._get_default_role_permissions(user_role)
-        all_permissions = list(role_permissions) + list(default_permissions)
-
-        # Special handling for command execution
-        if action_type == "allow_shell_execute":
-            return self._check_shell_execute_permission(all_permissions)
-
-        # Direct permission check
-        if action_type in role_permissions:
-            return True
-
-        # Check for wildcard permissions
-        if self._check_wildcard_permissions(action_type, role_permissions):
-            return True
-
-        # Check default role permissions
-        if action_type in default_permissions:
-            return True
-
-        logger.warning(f"Permission DENIED for role '{user_role}' to perform " f"action '{action_type}'")
-        return False
-
-    def _create_permission_denied_result(self, command: str, user: str, user_role: str) -> Dict[str, Any]:
-        """
-        Create a permission denied result for command execution.
-
-        Issue #281: Extracted from execute_command to reduce function length
-        and improve reusability.
-
-        Args:
-            command: The command that was attempted
-            user: Username who attempted execution
-            user_role: Role of the user
-
-        Returns:
-            Dict with error status and permission denied details
-        """
-        self.audit_log(
-            action="command_execution_denied",
-            user=user,
-            outcome="denied",
-            details={
-                "command": command,
-                "reason": "no_permission",
-                "role": user_role,
-            },
-        )
-        return {
-            "stdout": "",
-            "stderr": ("Permission denied: You do not have shell execution privileges"),
-            "return_code": 1,
-            "status": "error",
-            "security": {"blocked": True, "reason": "no_permission"},
-        }
-
-    def _should_force_approval(self, command: str, user_role: str) -> bool:
-        """
-        Determine if command requires forced approval based on role and risk.
-
-        Issue #281: Extracted from execute_command to reduce function length
-        and improve readability of approval logic.
-
-        Args:
-            command: Command to check
-            user_role: Role of user executing
-
-        Returns:
-            True if approval should be forced, False otherwise
-        """
-        role_permissions = self.roles.get(user_role, {}).get("permissions", [])
-        if not role_permissions:
-            # Fall back to default permissions if no configured role
-            role_permissions = self._get_default_role_permissions(user_role)
-
-        if "allow_all" in role_permissions:
-            return False
-
-        # Check if user has limited shell execution permissions
-        risk, _ = self.command_executor.assess_command_risk(command)
-        if "allow_shell_execute_safe" in role_permissions and risk != CommandRisk.SAFE:
-            # User can only execute safe commands without approval
-            return True
-        elif "allow_shell_execute" in role_permissions and risk in HIGH_RISK_COMMAND_RISKS:
-            # User has shell execute permission but high-risk commands need approval
-            return True
-
-        return False
-
-    def _get_default_role_permissions(self, user_role: str) -> List[str]:
-        """
-        Get default permissions for common user roles
-        SECURITY FIX: Removed god/superuser/root roles and granular admin permissions
-        """
-        default_role_permissions = {
-            # SECURITY FIX: Admin has elevated permissions but NOT unrestricted access
-            # Dangerous commands still require approval and validation for ALL users
-            "admin": [
-                "files.*",
-                "allow_goal_submission",
-                "allow_kb_read",
-                "allow_kb_write",
-                "allow_shell_execute",  # Can execute commands with proper validation
-                "allow_shell_high_risk",  # Can execute high-risk with approval
-                # NOTE: Dangerous commands ALWAYS require approval, even for admin
-            ],
-            "operator": [
-                "files.*",
-                "allow_goal_submission",
-                "allow_kb_read",
-                "allow_kb_write",
-                "allow_shell_execute",  # Standard shell execution
-                "allow_shell_moderate",  # Moderate risk commands with approval
-            ],
-            "developer": [
-                "files.*",
-                "allow_goal_submission",
-                "allow_kb_read",
-                "allow_kb_write",
-                "allow_shell_execute_safe",  # Only safe commands
-            ],
-            "user": [
-                "files.view",
-                "files.download",
-                "allow_goal_submission",
-                "allow_kb_read",
-            ],
-            "readonly": ["files.view", "files.download", "allow_kb_read"],
-            # Issue #744: Guest role REMOVED - security vulnerability
-            # Unauthenticated requests must be rejected, not assigned permissions
-        }
-
-        return default_role_permissions.get(user_role, [])
-
-    def _log_command_attempt(self, command: str, user: str, user_role: str, force_approval: bool) -> None:
-        """Log command execution attempt to audit log. Issue #620."""
-        self.audit_log(
-            action="command_execution_attempt",
-            user=user,
-            outcome="pending",
-            details={
-                "command": command,
-                "role": user_role,
-                "force_approval": force_approval,
-            },
-        )
-
-    async def _execute_basic_command(self, command: str) -> Dict[str, Any]:
-        """Execute command without security controls (fallback mode). Issue #620."""
-        process = await asyncio.create_subprocess_shell(
-            command, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE
-        )
-        stdout, stderr = await process.communicate()
-        return {
-            "stdout": stdout.decode().strip(),
-            "stderr": stderr.decode().strip(),
-            "return_code": process.returncode,
-            "status": "success" if process.returncode == 0 else "error",
-            "security": {"enabled": False},
-        }
-
-    def _log_command_complete(self, command: str, user: str, result: Dict[str, Any]) -> None:
-        """Log command execution completion to audit log. Issue #620."""
-        self.audit_log(
-            action="command_execution_complete",
-            user=user,
-            outcome=result["status"],
-            details={
-                "command": command,
-                "return_code": result["return_code"],
-                "security": result.get("security", {}),
-            },
-        )
-
-    async def execute_command(self, command: str, user: str, user_role: str) -> Dict[str, Any]:
-        """
-        Execute a command with security checks and audit logging
-
-        Args:
-            command: Command to execute
-            user: Username executing the command
-            user_role: Role of the user
-
-        Returns:
-            Execution result with security information
-        """
-        if not self.check_permission(user_role, "allow_shell_execute"):
-            return self._create_permission_denied_result(command, user, user_role)
-
-        force_approval = self._should_force_approval(command, user_role)
-        self._log_command_attempt(command, user, user_role, force_approval)
-
-        if self.enable_command_security:
-            result = await self.command_executor.run_shell_command(command, force_approval=force_approval)
-        else:
-            result = await self._execute_basic_command(command)
-
-        self._log_command_complete(command, user, result)
-        return result
-
-    def audit_log(self, action: str, user: str, outcome: str, details: Dict[str, Any]):
-        """Enhanced audit logging with command execution tracking"""
-        log_entry = {
-            "timestamp": datetime.datetime.now(tz=timezone.utc).isoformat(),
-            "user": user,
-            "action": action,
-            "outcome": outcome,
-            "details": details,
-        }
-
-        try:
-            with open(self.audit_log_file, "a", encoding="utf-8") as f:
-                f.write(json.dumps(log_entry) + "\n")
-            logger.debug("Audit log: %s by %s - %s", action, user, outcome)
-        except Exception as e:
-            logger.error("Failed to write to audit log: %s", e)
-
-    def get_command_history(self, user: str | None = None, limit: int = 100) -> List[Dict[str, Any]]:
-        """
-        Get command execution history from audit log
-
-        Args:
-            user: Filter by specific user (optional)
-            limit: Maximum number of entries to return
-
-        Returns:
-            List of command execution entries
-        """
-        command_history = []
-
-        try:
-            with open(self.audit_log_file, "r", encoding="utf-8") as f:
-                for line in f:
-                    # Use helper for parsing and filtering (Issue #315: depth reduction)
-                    entry = _parse_audit_log_entry(line, user)
-                    if entry:
-                        command_history.append(entry)
-        except FileNotFoundError:
-            return []
-
-        # Return most recent entries
-        return command_history[-limit:]
-
-    def authenticate_user(self, username: str, password: str) -> str | None:
-        """Authenticate user and return their role"""
-        if not self.enable_auth:
-            return "admin"
-
-        if username in self.allowed_users and self.allowed_users[username] == password:
-            # Map users to roles (simplified for demo)
-            user_roles = self.security_config.get("user_roles", {})
-            return user_roles.get(username, "user")
-
-        return None
-
-    def get_pending_approvals(self) -> List[Dict[str, Any]]:
-        """Get list of commands pending approval"""
-        pending = []
-        for cmd_id in self.pending_approvals:
-            pending.append(
-                {
-                    "command_id": cmd_id,
-                    "timestamp": cmd_id.split("_")[1] if "_" in cmd_id else "unknown",
-                }
-            )
-        return pending
-
-
-# Test the enhanced security layer
-if __name__ == "__main__":
-    logging.basicConfig(level=logging.INFO, format="%(message)s")
-
-    async def test_security():
-        """Test security layer with various commands and user roles."""
-        logger.info("Testing Enhanced Security Layer")
-        logger.info("%s", "=" * 60)
-
-        # Create enhanced security layer
-        security = EnhancedSecurityLayer()
-
-        # Test command risk assessment
-        test_commands = [
-            ("echo 'Hello World'", "admin"),
-            ("rm -rf /tmp/test", "admin"),
-            ("sudo apt update", "user"),
-            ("ls -la", "developer"),
-            ("cat /etc/passwd", "admin"),
-        ]
-
-        for cmd, role in test_commands:
-            logger.info(f"\nTesting: {cmd} (as {role})")
-            result = await security.execute_command(cmd, f"{role}_user", role)
-            logger.info(f"Result: {result['status']}")
-            if result.get("security"):
-                logger.info(f"Security: {result['security']}")
-            if result["stderr"]:
-                logger.info(f"Error: {result['stderr']}")
-            if result["stdout"]:
-                logger.info(f"Output: {result['stdout'][:100]}...")
-
-        # Show command history
-        logger.info("%s", "\n" + "=" * 60)
-        logger.info("Command History:")
-        history = security.get_command_history(limit=10)
-        for entry in history:
-            logger.info(f"- {entry['timestamp']}: {entry['user']} - " f"{entry['action']} - {entry['outcome']}")
-            if "command" in entry.get("details", {}):
-                logger.info(f"  Command: {entry['details']['command']}")
-
-    run_or_schedule(test_security())
